+++ qualang_tools/results/data_handler/data_processors/data_processor.py
@@ -1,7 +1,6 @@
 from abc import ABC
 from pathlib import Path
 
-
 __all__ = ["DataProcessor"]